Add form elements to the new attribute panel of the metadata form
adding the form text input and button is easy (we do that in the SpecifierComponent.vue
which now also has unit tests for it).

The interesting aspect is to ensure the form knows which "Attributes" button
was clicked and pass on the metadata (at the moment just the name) related to the corresponding file.

The toggleDrawer() method on the AnnotatorComponent.vue is how we achieve that.

We also need to make sure we can get reference to the drawer component [1]
At the moment it helps with the testing, but it may be helpful for future business logic too.

[1] https://vuejs.org/v2/api/#ref