Implement the form logic for entering file attributes for uploaded files
The collapsible panel on the metadata form that shows a new attribute form
now has the logic to collect the fields values and append it to
the list of attributes below the form.

unit tests was tricky, took some time to figure that we need to use ``mount`` instead of ``shallowMount``
so that the nested ElementUI's ``el-table`` component can be rendered in the test
and have expectation run against it.