Add FUW endpoint and GigaDB client to retrieve attributes for an upload
In order for GigaDB's metadata form to load existing attributes for file uploads
we need to be able to retrieve the attributes from File Upload Wizard in a similar way
to the file uploads.

So we need an endpoint in FUW REST API to get the attributes
(AttributeController.php's ``index`` action needed to be configured and the related AttributeSearch.php created)

And we need a new method ``getAttributes`` in FileUploadService.php to connect to the above endpoint.