Use plain SQL for table creation and manipulation
CDbMigration was used to create tables before but it does not
allow you to create columns with the type that is identical to
that in the production database. For example, if you use the
string abstract DB type, it can only be converted to varchar(255)
and sometimes varchar(50). For this reason, plain SQL is used for
all schema changes now.

Creating new rows in tables will use CDbMigration insert(). This
code seems cleaner and easier to make changes than when using SQL
insert statements to add rows into tables.