Fix watcher script that wasn't detecting uploaded files sometimes
When a file was uploaded using the wizard's uploader widget, the background
watcher container service sometimes didn't detect the file being moved to
the final drop box location so the meta-data wasn't updated in the database.

Configured the watcher to use polling instead of just waiting for file-system change.

As a consequence, we can reduce the wait time in some of the acceptance tests as an attempt to
fix the issue.

Also the acceptance tests for attributes can now be re-factored to using UX driven steps for setting uploads.