Correct FUW endpoint (and GigaDB client) for adding multiple attributes
In File Upload Wizard API,

The Yii2 routing was not correct as it required a specific upload_id as parameters to
the url while the spreadsheet parser add the upload id to the array of attributes
which is can span more than one uploads.

Secondly, the custom ReplaceAction is not a suitable location for "adding" new attributes,
can create confusion.
The default ``create`` route is more appropriate but we need to customize the CreateAction action
so it can support multiple attributes creation.

I've also modified the ``addAttributes`` GigaDB client in FileUploadService.php to make use of the new route.
There was also a discrepancy in the formatting of the attributes data being posted and the attributes returned by the
spreadsheet parser. Fixed that by moving the "Attributes" array envelope into the client to the API as the
spreadsheet parser returned array doesn't add and shouldn't know anything about it (as the need for "Attributes" is idiosyncratic to
the Yii2 routing/controller workflow).

Given the flow:

Metadata spreadsheet --> Vue.js Component (AnnotatorComponent.vue) --> GigaDB controller action (FilesAnnotateAction.php) --> Business object (DatasetUpload.php) --> GigaDB client to FUW API (FileUploadService.php) --> FUW controller action (CreateAction.php/ReplaceAction.php)

this commit deals with:

* GigaDB client to FUW API (FileUploadService.php)
* FUW controller action (CreateAction.php/ReplaceAction.php)