Add support for GigaDb file types in File Upload Wizard components
* DatasetUpload.php and DatasetUploadTest.php:

We've added the logic to match the data type in spreadsheet rows
with the reference data of GigaDB file types, which is sourced from a JSON file (which is created using the Yii command written in commit 28c9d274986c80e58ef0748908ac92966adc2e74).
A precise error is added to the error output array and that particular row is discarded.

* FilesAnnotateAction.php (controller),  filesAnnotate.php (view) and AuthorisedDatasetFilesAnnotateActionTest.php (functional test):

 Updated the controller to:
  * pass on the list of file types to the the form meta-data's javascript component (AnnotatorComponent.vue)
  * avoid triggering call to FUW API for updating uploads if the uploads array is empty (because they are errors preventing the upload)
  * avoid triggering call to FUW API for updating attributes if the attributes array is empty
  * cosmetic fix to the error message so they are more readable

* AnnotatorComponent.vue *(Vue.js component), AnnotatorComponent.spec.js (unit test) and db.json (test data):

The component receives the list of GigaDB file types as props in JSON format.

* Acceptance tests files *.feature (stories and scenarios), MetadataFormSteps.php (steps implementations) and sample*.csv (test data):

Implemented acceptance test scenarios for handling of errors related to data types and upgrade test data for all the FUW acceptance scenarios
to use GigaDB file types.

* bulk-data-upload-example.csv

example of meta-data spreadsheet for authors, also used as test data in AuthorisedDatasetFilesAnnotateActionTest.php,
It's now updated to use GigaDB file type.