Add initial work on back-end for the mock-up page for reviewers
For this piece, we will leverage the SOLID Russian doll arrangement with have in DatasetController
(Formatted* -> Cached* -> Stored*) where we will
create a new type of DatasetComponent (ResourcedDatasetFiles.php) that will replace StoredDatasetFiles.php
when the dataset view page is accessed for review purpose.

This commit as a basic (no support for sample id yet) working version of the DAO method to get the list of files, converting the File Upload Wizard file format
into the GigaDB file format along the way.