Implement editors setting dataset upload status to Rejected/Submitted
The main business object for file upload workflow, DatasetUpload.php has been updated to add
a method to set dataset upload status to "Submitted".
That methods also include sending an email to database@gigasciencejournal.com and adding
an  entry in the curation log.

That required a new email template for the "Submitted" status (``files/templates/Submitted.twig``)
and additional configuration to Yii2 backend of GigaDB to configure the new email address to send that notification to.

AdminDatasetController was updated to detect if the upload status has been changed to "Submitted" in which case we invoke
the business logic for the upload workflow.

The acceptance scenario has a step checking for email being sent. Implementing that step (in ``CuratorSteps.php``) required installing additional libraries for
manipulating emails, hence the change to ``composer.json`` and required the use of PHPunit's type of assertions, hence the change to ``acceptance.suite.yml``.