Show file attributes associated with file uploads on dataset mockup page
The main change was done in ``ResourcedDatasetFiles.php`` to collect attributes (by querying the FUW API using an instance of ```FileUploadService.php``) into the the same key already used in the structure found in ``StoredDatasetFiles.php``.

By virtue of keeping the same structure and implementing the ``DatasetFilesInterface.php`` interface,
``ResourcedDatasetFiles.php`` can be used interchangeably within the S.O.L.I.D Russian doll pattern of StoredDatasetFiles|ResourcedDatasetFiels < CachedDatasetFiles < FormattedDatasetFiles,
thus allowing the mockup page and the regular Dataset view to share most of their code.

In order for the acceptance scenario to play the steps proving the functionality I had to change the display default for the files tables
so that the files attributes are visible by default (see ``DatasetPageSettings.php`` and its test file).

The steps implementations for the new acceptance scenario (in ``file-upload-6-reviewer-access-drop-box-files.feature``)
are in ``ReviewerSteps.php``.