Show samples associated with file uploads on dataset mockup page
he main change was done in ``ResourcedDatasetFiles.php`` to convert the content of ``sample_ids`` attribute
of the ``upload`` model into the structure expected by the formatting function ``formatDatasetFilesSamples()`` from ``FormattedDatasetFiles.php``.

By virtue of keeping the same structure and implementing the ``DatasetFilesInterface.php`` interface,
``ResourcedDatasetFiles.php`` can be used interchangeably within the S.O.L.I.D Russian doll pattern of StoredDatasetFiles|ResourcedDatasetFiels < CachedDatasetFiles < FormattedDatasetFiles,
thus allowing the mockup page and the regular Dataset view to share most of their code.

Prior to that I had to change the upload table in FUW to have a ``sample_ids`` string column instead of a ``sample_id`` integer column
(hence the change to the migration file and to the ``Upload.php`` model class.

Fixed the functional tests ``DatasetMockupViewActionTest.php`` whose test data was no longer correct (mockupUrl record need to be accompanied by
a FUW credentials records for the reviewer for which the mockup URL is being generated for).

The acceptance test scenario for this feature was straightforward, just needed a new step implementation with the required test data.