Implement the JS side of MD5 checksum of uploaded files
Now, when an author uploads files in File Upload Wizard's web interface (that uses the Vue.js application)
the MD5 checksum is calculated as file metadata.

This is done by writing a custom plugin "Checksum" ( in ``uppy-checksum.js``) to the Uppy file uploader
that will add a pre-processor hook to the upload pipeline that will calculate MD5 checksum
using the Javascript libraries ``spark-md5`` and ``chunked-file-reader`` (dependencies added in ``package.json``)

The plugin is configured in ``UploaderComponent.vue`` and the test of the MD5 calculation process
has been added to ``UploaderComponent.spec.js``

To load the test data, we need to use the Webpack plugin ``raw-loader`` hence the addition
to the Webpack config ``webpack.config.js`` and a new dependency in ``package.json``.

The plugin was made following the guidelines from Uppy documentation on writing plugins:

* https://uppy.io/docs/writing-plugins/

This is only half the work, next we still need to extract and store the MD5 checksum server-side.