Prevent meta-data form from being submitted when new sample is attached
When entering a new sample in the sample field of the panel floating over the metadata form,
the "Complete and return to Your Uploaded Datasets page" button to submit the metadata
to the server was picking up the enter/return key event and prematurely submitted the form.

This is because the input field in ``SamplerComponent.vue`` subscribed to "keyup" DOM (native) event
to detect that the return/enter key has been pressed and released, but not the "keydown" event
which was eventually picked up by the "submit" button "Complete and return to Your Uploaded Datasets page".

The fix is for the new sample input field to listen to "keydown" event and stops its propagation to other listener.
hence the corrected listener registration "@keydown.enter.native.stop" on the new sample input field.