Fix Sample IDs UI so newly added samples are attached to the upload
The Sample UI component (``SamplerComponent.vue``) wasn't passing back
to the parent meta-data form component ( ``AnnotatorComponent.vue``) the newly added samples.

This is because when no samples for an upload exist yet, the collection attribute to <sampler>
used to pass props to the component is undefined, so there's no reference.

Instead, we pass the new sample as a parameter to the event emitted when user clicks on the Save button
in the Sample UI component.
The event handler in ``AnnotatorComponent.vue`` ("setSampleIds()") will process the new parameter
and stringify its array to assign to upload's sample_ids property.
We still need to maintain an array of the samples, so the user can add new samples several times without starting from scratch.

Also added in the confirmation message after completing the FUW workflow the number of samples added (see ``FilesAnnotateAction.php``)
and updated acceptance scenario to test the sample UI workflow.